output public url to artifacts; use localIdentifier for react-component
